refactor(clock): HLE-54 move SystemClock/FrozenClock to adapters/clock
Concrete ClockPort implementations belong in adapters, not core. Move
SystemClock + FrozenClock to atman.adapters.clock and update external
importers (tests, e2e, demos) to import from the new location.

Core services that defaulted to SystemClock() in __init__ now defer the
import to runtime inside the constructor (executed only when no clock is
supplied) so the Core module itself has no static dependency on a
concrete adapter. core/clock_impl.py keeps only ensure_utc — a pure
timezone-normalisation helper that has no implementation dependency.
