+++ e2e/full_loop.py
@@ -42,7 +42,7 @@ from atman.adapters.storage.in_memory_reflection_store import (
     InMemoryPatternStore,
     InMemoryReflectionEventStore,
 )
-from atman.core.clock_impl import FrozenClock, SystemClock
+from atman.adapters.clock import FrozenClock, SystemClock
 from atman.core.models import (
     CoreValue,
     Goal,
