+++ e2e/scenarios/session_lifecycle_timeout.py
@@ -35,7 +35,7 @@ from pathlib import Path
 from uuid import uuid4
 
 from atman.adapters.storage.file_state_store import FileStateStore
-from atman.core.clock_impl import SystemClock
+from atman.adapters.clock import SystemClock
 from atman.core.models import (
     CoreValue,
     Goal,
