+++ tests/test_domain_invariants.py
@@ -46,7 +46,7 @@ import pytest
 
 from atman.adapters.memory import InMemoryBackend
 from atman.adapters.storage import FileStateStore, InMemoryStateStore
-from atman.core.clock_impl import FrozenClock
+from atman.adapters.clock import FrozenClock
 from atman.core.models import (
     CoreValue,
     EmotionalDepth,
