+++ tests/test_entity_stance_formulator.py
@@ -10,7 +10,7 @@ from atman.adapters.memory.in_memory_entity_registry import InMemoryEntityRegist
 from atman.adapters.memory.in_memory_entity_stance import InMemoryEntityStanceStore
 from atman.adapters.reflection.mock_reflection_model import MockReflectionModel
 from atman.adapters.storage.in_memory_state_store import InMemoryStateStore
-from atman.core.clock_impl import FrozenClock
+from atman.adapters.clock import FrozenClock
 from atman.core.models.entity import Entity, EntityType
 from atman.core.models.experience import EmotionalDepth, FeltSense, KeyMoment
 from atman.core.models.reflection import StanceFormulationOutput
